Use the same copy logic for primitive slices as for the complex.

The opposite approach of https://github.com/open-telemetry/opentelemetry-collector/pull/13293. 

Apparently, there is slight improvement when 10-items slice is copied into 7-items slice.

Not sure if we want to complicate the logic for this insignificant gain.

before 
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/pdata/internal
cpu: Apple M1 Max
BenchmarkCopyOrigInt64Slice/0_to_7-10         	 8608524	       139.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkCopyOrigInt64Slice/1_to_7-10         	 8533219	       140.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkCopyOrigInt64Slice/7_to_7-10         	 8768828	       137.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkCopyOrigInt64Slice/10_to_7-10        	 7211601	       164.7 ns/op	     112 B/op	       1 allocs/op
BenchmarkCopyOrigInt64Slice/20_to_7-10        	 7005366	       171.8 ns/op	     160 B/op	       1 allocs/op
BenchmarkCopyOrigInt64Slice/50_to_7-10        	 5943145	       209.0 ns/op	     416 B/op	       1 allocs/op
```

after
```
BenchmarkCopyOrigInt64Slice/0_to_7-10         	 8667313	       140.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkCopyOrigInt64Slice/1_to_7-10         	 8764303	       136.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkCopyOrigInt64Slice/7_to_7-10         	 8669146	       139.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkCopyOrigInt64Slice/10_to_7-10        	 7425303	       160.8 ns/op	      80 B/op	       1 allocs/op
BenchmarkCopyOrigInt64Slice/20_to_7-10        	 6743520	       177.6 ns/op	     160 B/op	       1 allocs/op
BenchmarkCopyOrigInt64Slice/50_to_7-10        	 5643139	       213.9 ns/op	     416 B/op	       1 allocs/op
```